fix(mcp): build checkout preview purchase plan
