+++ internal/cli/commands_checkout.go
@@ -1,18 +1,13 @@
 package cli
 
 import (
-	"context"
-	"fmt"
-	"regexp"
 	"strings"
 
-	"github.com/mekedron/wolt-cli/internal/domain"
 	woltgateway "github.com/mekedron/wolt-cli/internal/gateway/wolt"
+	"github.com/mekedron/wolt-cli/internal/service/checkoutpayload"
 	"github.com/mekedron/wolt-cli/internal/service/output"
 	"github.com/spf13/cobra"
 )
-
-var objectIDPattern = regexp.MustCompile(`^[a-f0-9]{24}$`)
 
 func newCheckoutCommand(deps Dependencies) *cobra.Command {
 	checkout := newCheckoutPreviewCommand(deps)
@@ -99,9 +94,9 @@ func newCheckoutPreviewCommand(deps Dependencies) *cobra.Command {
 				)
 			}
 
-			checkoutPayload, checkoutWarnings, err := buildCheckoutPayload(
+			checkoutPayload, checkoutWarnings, err := checkoutpayload.Build(
 				cmd.Context(),
-				deps,
+				deps.Wolt,
 				basket,
 				location,
 				deliveryMode,
@@ -187,374 +182,6 @@ func newCheckoutPreviewCommand(deps Dependencies) *cobra.Command {
 		lonSet = cmd.Flags().Changed("lon")
 	}
 	return cmd
-}
-
-func buildCheckoutPayload(
-	ctx context.Context,
-	deps Dependencies,
-	basket map[string]any,
-	location domain.Location,
-	deliveryMode string,
-	tip int,
-	promoCode string,
-) (map[string]any, []string, error) {
-	deliveryMode = strings.ToLower(strings.TrimSpace(deliveryMode))
-	if deliveryMode == "" {
-		deliveryMode = "standard"
-	}
-	if deliveryMode != "standard" && deliveryMode != "priority" && deliveryMode != "schedule" {
-		return nil, nil, fmt.Errorf("unsupported --delivery-mode %q", deliveryMode)
-	}
-
-	venue := asMap(basket["venue"])
-	venueID := strings.TrimSpace(asString(venue["id"]))
-	currency := inferCurrency(asString(basket["total"]))
-	if currency == "" {
-		currency = "EUR"
-	}
-	country := strings.TrimSpace(asString(venue["country"]))
-	warnings := []string{}
-	itemDetails := map[string]map[string]any{}
-	categoryIDsByItemID := map[string]string{}
-	assortmentPayload := map[string]any{}
-
-	venueSlug := resolveBasketVenueSlug(venue)
-	if venueSlug != "" && deps.Wolt != nil {
-		if payload, err := deps.Wolt.AssortmentByVenueSlug(ctx, venueSlug); err == nil {
-			assortmentPayload = payload
-			mergeCheckoutCategoryIndexes(categoryIDsByItemID, buildCheckoutCategoryIDIndex(payload))
-		} else {
-			warnings = append(warnings, fmt.Sprintf("unable to load venue assortment payload for category mapping (slug=%s)", venueSlug))
-		}
-		if payload, err := cachedVenuePageStatic(ctx, deps, venueSlug); err == nil {
-			mergeCheckoutCategoryIndexes(categoryIDsByItemID, buildCheckoutCategoryIDIndex(payload))
-		}
-	}
-
-	menuItems := make([]any, 0, len(asSlice(basket["items"])))
-	for _, value := range asSlice(basket["items"]) {
-		item := asMap(value)
-		itemID := strings.TrimSpace(asString(item["id"]))
-		count := asInt(item["count"])
-		if count <= 0 {
-			count = 1
-		}
-		price := asInt(item["price"])
-		if price <= 0 {
-			return nil, warnings, fmt.Errorf("unable to resolve base_price for basket item %q", itemID)
-		}
-
-		detail := map[string]any{}
-		if itemID != "" && deps.Wolt != nil {
-			if cached, ok := itemDetails[itemID]; ok {
-				detail = cached
-			} else if payload, err := deps.Wolt.VenueItemPage(ctx, venueID, itemID); err == nil {
-				detail = payload
-				itemDetails[itemID] = payload
-				mergeCheckoutCategoryIndexes(categoryIDsByItemID, buildCheckoutCategoryIDIndex(payload))
-			} else if len(assortmentPayload) > 0 {
-				detail = assortmentPayload
-				itemDetails[itemID] = assortmentPayload
-			} else {
-				warnings = append(warnings, fmt.Sprintf("unable to enrich checkout payload for item %s; using basket defaults", itemID))
-			}
-		}
-
-		categoryID := resolveCheckoutCategoryID(item, detail, itemID, categoryIDsByItemID)
-		if categoryID == "" {
-			if looksLikeObjectID(itemID) {
-				categoryID = itemID
-				warnings = append(warnings, fmt.Sprintf("unable to resolve category_id for item %s; falling back to item id", itemID))
-			} else {
-				return nil, warnings, fmt.Errorf("unable to resolve category_id for basket item %q", itemID)
-			}
-		}
-		categoryIDs := resolveCheckoutCategoryIDs(item, categoryID)
-		valuePrices := buildOptionValuePriceIndex(detail)
-		options := buildCheckoutOptions(item["options"], valuePrices)
-
-		menuItems = append(menuItems, map[string]any{
-			"id":                                itemID,
-			"venue_id":                          venueID,
-			"count":                             count,
-			"base_price":                        price,
-			"end_amount":                        count * price,
-			"is_weighted_item":                  false,
-			"category_id":                       categoryID,
-			"category_ids":                      categoryIDs,
-			"alcohol_permille":                  asInt(coalesceAny(item["alcohol_permille"], 0)),
-			"exclude_from_credits":              asBool(coalesceAny(item["exclude_from_credits"], false)),
-			"exclude_from_discounts":            asBool(coalesceAny(item["exclude_from_discounts"], false)),
-			"exclude_from_discounts_min_basket": asBool(coalesceAny(item["exclude_from_discounts_min_basket"], false)),
-			"restrictions":                      coalesceAny(item["restrictions"], []any{}),
-			"age_limit":                         coalesceAny(item["age_limit"], nil),
-			"options":                           options,
-		})
-	}
-
-	promoDiscountIDs := []any{}
-	if strings.TrimSpace(promoCode) != "" {
-		promoDiscountIDs = append(promoDiscountIDs, strings.TrimSpace(promoCode))
-	}
-
-	return map[string]any{
-		"purchase_plan": map[string]any{
-			"venue": map[string]any{
-				"id":       venueID,
-				"currency": currency,
-				"country":  country,
-			},
-			"delivery_method":           "homedelivery",
-			"menu_items":                menuItems,
-			"use_promo_discount_ids":    promoDiscountIDs,
-			"courier_tip":               tip,
-			"use_cash":                  false,
-			"use_credits_and_tokens":    false,
-			"use_loyalty_points_amount": 0,
-			"use_promo_surcharge_ids":   []any{},
-			"payment_methods":           []any{},
-			"is_priority_delivery":      deliveryMode == "priority",
-			"delivery": map[string]any{
-				"delivery_coordinates": map[string]any{
-					"latitude":  location.Lat,
-					"longitude": location.Lon,
-				},
-			},
-		},
-	}, warnings, nil
-}
-
-func resolveCheckoutCategoryID(item map[string]any, detail map[string]any, itemID string, fallback map[string]string) string {
-	if id := strings.TrimSpace(asString(item["category_id"])); id != "" {
-		return id
-	}
-	if category := asMap(item["category"]); category != nil {
-		if id := strings.TrimSpace(asString(coalesceAny(category["id"], category["_id"]))); id != "" {
-			return id
-		}
-	}
-	if categoryIDs := asSlice(item["category_ids"]); len(categoryIDs) > 0 {
-		if id := strings.TrimSpace(asString(categoryIDs[0])); id != "" {
-			return id
-		}
-	}
-	if detailCategory := resolveCheckoutCategoryIDFromItemLikePayload(detail); detailCategory != "" {
-		return detailCategory
-	}
-	if id := resolveCheckoutCategoryIDFromDetails(detail, itemID); id != "" {
-		return id
-	}
-	if id := strings.TrimSpace(fallback[itemID]); id != "" {
-		return id
-	}
-	return ""
-}
-
-func resolveCheckoutCategoryIDFromDetails(detail map[string]any, itemID string) string {
-	if strings.TrimSpace(itemID) == "" {
-		return ""
-	}
-	categoryIDsByItemID := buildCheckoutCategoryIDIndex(detail)
-	if id := strings.TrimSpace(categoryIDsByItemID[itemID]); id != "" {
-		return id
-	}
-	return ""
-}
-
-func resolveCheckoutCategoryIDFromItemLikePayload(payload map[string]any) string {
-	if payload == nil {
-		return ""
-	}
-	if id := strings.TrimSpace(asString(payload["category_id"])); id != "" {
-		return id
-	}
-	if category := asMap(payload["category"]); category != nil {
-		if id := strings.TrimSpace(asString(coalesceAny(category["id"], category["_id"]))); id != "" {
-			return id
-		}
-	}
-	if categoryIDs := asSlice(payload["category_ids"]); len(categoryIDs) > 0 {
-		if id := strings.TrimSpace(asString(categoryIDs[0])); id != "" {
-			return id
-		}
-	}
-	return ""
-}
-
-func resolveBasketVenueSlug(venue map[string]any) string {
-	if venue == nil {
-		return ""
-	}
-	candidates := []any{
-		venue["slug"],
-		venue["venue_slug"],
-		venue["public_slug"],
-		venue["url_slug"],
-	}
-	for _, candidate := range candidates {
-		if slug := strings.TrimSpace(asString(candidate)); slug != "" {
-			return slug
-		}
-	}
-	return ""
-}
-
-func buildCheckoutCategoryIDIndex(payload map[string]any) map[string]string {
-	index := map[string]string{}
-	if payload == nil {
-		return index
-	}
-	var walk func(any)
-	walk = func(node any) {
-		switch value := node.(type) {
-		case map[string]any:
-			if categories := asSlice(value["categories"]); len(categories) > 0 {
-				for _, categoryNode := range categories {
-					collectCheckoutCategoryMappings(categoryNode, index)
-				}
-			}
-			if menuItems := asSlice(value["menu_items"]); len(menuItems) > 0 {
-				for _, menuItemNode := range menuItems {
-					menuItem := asMap(menuItemNode)
-					if menuItem == nil {
-						continue
-					}
-					itemID := strings.TrimSpace(asString(coalesceAny(menuItem["item_id"], menuItem["id"])))
-					if itemID == "" {
-						continue
-					}
-					if categoryID := resolveCheckoutCategoryIDFromItemLikePayload(menuItem); categoryID != "" {
-						index[itemID] = categoryID
-					}
-				}
-			}
-			collectCheckoutCategoryMappings(value, index)
-			for _, nested := range value {
-				walk(nested)
-			}
-		case []any:
-			for _, nested := range value {
-				walk(nested)
-			}
-		}
-	}
-	walk(payload)
-	return index
-}
-
-func collectCheckoutCategoryMappings(node any, index map[string]string) {
-	category := asMap(node)
-	if category == nil {
-		return
-	}
-	categoryID := strings.TrimSpace(asString(coalesceAny(category["category_id"], category["id"], category["_id"])))
-	if categoryID == "" {
-		return
-	}
-	for _, itemNode := range asSlice(category["item_ids"]) {
-		itemID := strings.TrimSpace(asString(itemNode))
-		if itemID == "" {
-			continue
-		}
-		index[itemID] = categoryID
-	}
-	for _, itemNode := range asSlice(category["items"]) {
-		itemID := strings.TrimSpace(asString(itemNode))
-		if item := asMap(itemNode); item != nil {
-			itemID = strings.TrimSpace(asString(coalesceAny(item["item_id"], item["id"])))
-		}
-		if itemID == "" {
-			continue
-		}
-		index[itemID] = categoryID
-	}
-}
-
-func mergeCheckoutCategoryIndexes(target map[string]string, source map[string]string) {
-	if target == nil || len(source) == 0 {
-		return
-	}
-	for itemID, categoryID := range source {
-		itemID = strings.TrimSpace(itemID)
-		categoryID = strings.TrimSpace(categoryID)
-		if itemID == "" || categoryID == "" {
-			continue
-		}
-		if _, exists := target[itemID]; exists {
-			continue
-		}
-		target[itemID] = categoryID
-	}
-}
-
-func looksLikeObjectID(value string) bool {
-	return objectIDPattern.MatchString(strings.ToLower(strings.TrimSpace(value)))
-}
-
-func resolveCheckoutCategoryIDs(item map[string]any, categoryID string) []any {
-	categoryIDs := asSlice(item["category_ids"])
-	if len(categoryIDs) > 0 {
-		return categoryIDs
-	}
-	if strings.TrimSpace(categoryID) == "" {
-		return []any{}
-	}
-	return []any{categoryID}
-}
-
-func buildOptionValuePriceIndex(detail map[string]any) map[string]int {
-	index := map[string]int{}
-	for _, spec := range extractOptionSpecs(detail) {
-		for valueID, value := range spec.Values {
-			valueID = strings.TrimSpace(valueID)
-			if valueID == "" {
-				continue
-			}
-			index[valueID] = value.Price
-		}
-	}
-	return index
-}
-
-func buildCheckoutOptions(raw any, valuePrices map[string]int) []any {
-	options := make([]any, 0, len(asSlice(raw)))
-	for _, optionValue := range asSlice(raw) {
-		option := asMap(optionValue)
-		if option == nil {
-			continue
-		}
-
-		values := make([]any, 0, len(asSlice(option["values"])))
-		for _, selectedValue := range asSlice(option["values"]) {
-			value := asMap(selectedValue)
-			if value == nil {
-				continue
-			}
-			valueID := strings.TrimSpace(asString(value["id"]))
-			if valueID == "" {
-				continue
-			}
-			count := asInt(value["count"])
-			if count <= 0 {
-				count = 1
-			}
-			price := asInt(value["price"])
-			if inferred, ok := valuePrices[valueID]; ok {
-				price = inferred
-			}
-			values = append(values, map[string]any{
-				"id":    valueID,
-				"count": count,
-				"price": price,
-			})
-		}
-
-		options = append(options, map[string]any{
-			"id":     strings.TrimSpace(asString(option["id"])),
-			"values": values,
-		})
-	}
-	return options
 }
 
 func findTotalFormattedAmount(payload map[string]any) string {
